Add sample_txs function to perf_utils shared crate
Shared code between bench-tps and bench-exchange
