+++ client/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod client_error;
 mod generic_rpc_client_request;
 pub mod mock_rpc_client_request;
+pub mod perf_utils;
 pub mod rpc_client;
 pub mod rpc_client_request;
 pub mod rpc_request;
